requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
beautifulsoup4==4.12.3
lxml==4.9.3
Pillow==10.3.0
ImageHash==4.3.1
aiofiles==23.2.1
python-multipart==0.0.9
numpy==1.26.4
